Add initial setup to SyncController and update config, sync model
In order to enable university-agnostic mode, sync controller was
updated to first show the user the setup wizard if they do not have
a predefined university, which is stored as an index.
After the wizard, or after the configuration file is read on the
next execution, the login window is constructed and then shown to
the user.
Note that at the moment it is not possible to sign out, since the
cookie store is kept out of the users control.
Sync model has been updated to hold the university data with help
from an updated config class, and a new helper university class.