Enable stopping a job mid-download
Closes #25

Before this change, if a user quit the program mid download, they
would have to wait until the download stopped completely for the
application to close. This change will let the user close the app
even in the middle of a download. Of course, the download will be
interrupted in the middle, but the last sync time wont be modified
so next time the items should redownload from previous complete sync