Add pydantic validation error handling and user notification
Previously, sync.py would consider any ValueError to be a credential
expiration error, and if this was instead a ValidationError from
pydantic, it would get stuck in a loop trying to progress further.

With this change, ValidationError is handled on its own, in which case
the sync task is stopped forever, and the user will get notified of this
the next time they activate the tray icon. A browser tab will also
be opened to report the issue on Github.

Furthermore, both ValidationError and ValueError are now also logged.
Finally, user notification msgs have been all contained in one dictionary.

Thanks to @cainanmahar for bringing this up.