Created new BBAvailable enum
Previously, the `available` field of a BBAvailability was allowed to
simply be a string, which prevented users from verifying the effective
boolean value of the `available` field. Now, the `available` field is
mapped to a BBAvailable enum, which has a boolean cast method. If that
enum does not contain the string that BBAvailability is passed on
construction, then BBAvailability will allow itself to contain a string
in the `available` field. Casting a string to a bool returns True unless
the string is empty, so the field will behave generally as expected.