Explicitly set pyqt5-qt5 version on Pipfile, update PyQt5
PyQt5 was updated due to a security release concerning a recent cURL
vulnerability. Similarly, pyqt5-qt5 offered a similar patch with
distribution only for macOS. Since the Pipfile had been locked on that
platform, this would lead to the Pipfile.lock not being valid on other
platforms, even when pip expected it to.
Similarly, the PyQt5 version was also bumped in pyproject.toml