Remove BBAvailable bool support and add Other member, add tests
BBAvailable and BBAvailability were refactored to remove any support for
boolean availability values, and a new member 'Other' was added to group
any unexpected values without throwing an error.

(Title case was used for the enum member names, following the convention
set by the API itself, and because it looks better.)

While BBAvailable itself can cast bool input values through its
`_missing_` method, pydantic does not seem to recognise this, which
although possible to correct with a 'before' validator does not seem to
be necessary. It is therefore assumed that the Blackboard API only
returns string values for this field.

Finally, some tests were added for BBAvailable.

Co-authored-by: Cainan Mahar <52504207+cainanmahar@users.noreply.github.com>