Fix issue due to package pyqt5-qt5 not having a universal version
To understand why this issue came to be, we first need to understand
the context.

1. GitHub decided to migrate the macos-latest runner image to macOS 14
   Nothing wrong with that by itself. However, macOS 14 has mainly arm64
   images. This is documented in actions/runner-images#9741
2. For some reason that baffles me, riverbank decided to ship their
   pyqt5-qt5 in a way that only includes certain platforms in each
   release. For instance, version X will only have wheels for windows,
   while version Y will have only a macOS build. Keep this in mind.
3. Up to a few months, builds were perfectly fine.

However, once GitHub enforced the new version of macos-latest
two things happened:

a. Suddenly there was no valid Python 3.10.10 image for macos.
   This is because while there is a valid x86 image, there never was
   an arm64 one. I fixed this in a prior commit.
b. Suddenly there was no valid version of pyqt5-qt5. This is for the
   same reason.

I actually had let @dependabot update pyqt5-qt5 without realising that
the last 4 or so releases have not had a windows build in them.
But once I reverted this commit I realised the macOS build was broken.

Fair enough, now it's a matter of selecting a version equal or greater
to than the minimum common denominator, no?

But I haven't mentioned that Pipenv is extremely constrained in this
regard. So once it generates a Pipfile.lock, these requirements are
locked to the version it managed to resolve at that time. Which presents
a problem, since it will be dependant on the platform this step was
executed on.

I even tried using markers to select a version based on the os, but
two different versions cannot be specified at all, even if mutually
exclusive.

After all, I ended up manually replacing the requirement generated by
Pipenv so that the version is selected using >= rather than an exact
match.